+++ packages/test/src/activities/index.ts
@@ -1,5 +1,7 @@
+/* eslint-disable @typescript-eslint/no-empty-function */
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 import fetch from 'node-fetch';
-import { Context, CancellationError } from '@temporalio/activity';
+import { Context, CancelledError } from '@temporalio/activity';
 import { Connection } from '@temporalio/client';
 import { fakeProgress as fakeProgressInner } from './fake-progress';
 
@@ -21,6 +23,28 @@ export async function echo(message?: string): Promise<string> {
 export async function httpGet(url: string): Promise<string> {
   return `<html><body>hello from ${url}</body></html>`;
 }
+
+/**
+ * Just a mock, used in Workflow samples
+ */
+export async function httpGetJSON(url: string): Promise<any> {
+  return { url };
+}
+
+/**
+ * Just a mock, used in Workflow samples as an example of an activity that creates a side-effect
+ */
+export async function httpPostJSON(_url: string, _data: any): Promise<void> {}
+
+/**
+ * Mock for Workflow samples
+ */
+export async function setup(): Promise<void> {}
+
+/**
+ * Mock for Workflow samples, used to demo cleanup (e.g. after cancellation)
+ */
+export async function cleanup(_url: string): Promise<void> {}
 
 export async function throwAnError(message: string): Promise<void> {
   throw new Error(message);
@@ -45,7 +69,7 @@ export async function fakeProgress(sleepIntervalMs = 1000): Promise<void> {
   try {
     await fakeProgressInner(sleepIntervalMs);
   } catch (err) {
-    if (err instanceof CancellationError) {
+    if (err instanceof CancelledError) {
       try {
         await signalSchedulingWorkflow('activityCancelled');
       } catch (signalErr) {