+++ packages/test/src/workflows/cancel-http-request.ts
@@ -1,7 +1,6 @@
-import { Context, CancellationError, cancel, sleep } from '@temporalio/workflow';
+import { Context, CancellationScope, CancelledError, sleep, Trigger } from '@temporalio/workflow';
 import { CancellableHTTPRequest } from '../interfaces';
 import { cancellableFetch } from '@activities';
-import { ResolvablePromise } from './resolvable-promise';
 
 const fetch = Context.configure(cancellableFetch, {
   type: 'remote',
@@ -9,33 +8,35 @@ const fetch = Context.configure(cancellableFetch, {
   heartbeatTimeout: '3s',
 });
 
-const activityStarted = new ResolvablePromise<void>();
-const activityCancelled = new ResolvablePromise<void>();
+const activityStarted = new Trigger<void>();
+const activityCancelled = new Trigger<void>();
 
 const signals = {
   activityStarted(): void {
-    activityStarted.resolve(undefined);
+    activityStarted.resolve();
   },
   activityCancelled(): void {
-    activityCancelled.resolve(undefined);
+    activityCancelled.resolve();
   },
 };
 
 async function main(url: string, waitForActivityCancelled: boolean): Promise<void> {
-  const promise = fetch(url, true);
-  await activityStarted;
-  cancel(promise);
   try {
-    await promise;
+    await CancellationScope.cancellable(async () => {
+      const promise = fetch(url, true);
+      await activityStarted;
+      CancellationScope.current().cancel();
+      await promise;
+    });
   } catch (err) {
-    if (err instanceof CancellationError) {
+    if (err instanceof CancelledError) {
       if (waitForActivityCancelled) {
         await Promise.race([
           activityCancelled,
-          (async () => {
+          CancellationScope.nonCancellable(async () => {
             await sleep(10000);
             throw new Error('Confirmation of activity cancellation never arrived');
-          })(),
+          }),
         ]);
       }
     }