+++ packages/workflow/src/errors.ts
@@ -33,21 +33,13 @@ export class WorkflowExecutionContinuedAsNewError extends Error {
 }
 
 /**
- * Used to denote where the cancellation was originated
- *
- * - external - The workflow was cancelled by an external API call
- * - internal - Cancellation was requested by using `cancel` from within a workflow
- */
-export type CancellationSource = 'internal' | 'external';
-
-/**
  * Thrown in workflow when it is requested to be cancelled either externally or internally.
  *
  * @see {@link CancellationSource}
  */
-export class CancellationError extends Error {
-  public readonly name: string = 'CancellationError';
-  public constructor(message: string, public readonly source: CancellationSource) {
+export class CancelledError extends Error {
+  public readonly name: string = 'CancelledError';
+  public constructor(message: string) {
     super(message);
   }
 }