+++ packages/meta/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/temporalio/sdk-node/compare/temporalio@0.2.3...temporalio@0.2.4) (2021-06-18)
+
+**Note:** Version bump only for package temporalio
+
+
+
+
+
+## [0.2.3](https://github.com/temporalio/sdk-node/compare/temporalio@0.2.2...temporalio@0.2.3) (2021-06-18)
+
+**Note:** Version bump only for package temporalio
+
+
+
+
+
+## [0.2.2](https://github.com/temporalio/sdk-node/compare/temporalio@0.2.1...temporalio@0.2.2) (2021-06-16)
+
+**Note:** Version bump only for package temporalio
+
+
+
+
+
 ## [0.2.1](https://github.com/temporalio/sdk-node/compare/temporalio@0.2.0...temporalio@0.2.1) (2021-05-17)
 
 **Note:** Version bump only for package temporalio